requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
openai==1.14.0
python-dotenv==1.0.1